fix(android/app): Use BaseActivity
