+++ android/KMAPro/kMAPro/src/main/java/com/tavultesoft/kmapro/InfoActivity.java
@@ -20,7 +20,7 @@ import android.widget.TextView;
 import android.annotation.SuppressLint;
 import androidx.appcompat.app.AppCompatActivity;
 
-public class InfoActivity extends AppCompatActivity {
+public class InfoActivity extends BaseActivity {
   private final static String TAG = "InfoActivity";
   private WebView webView;
   private final String HELP_PRODUCTION_HOST = "help.keyman.com";
